backend/requirements.txt
# PolicyMitr Backend
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
supabase>=2.0.0
google-generativeai>=0.3.0
edge-tts>=6.1.0
PyMuPDF>=1.23.0
pytesseract>=0.3.10
Pillow>=10.0.0
python-multipart>=0.0.6
pydantic>=2.0.0
transformers>=4.36.0
torch>=2.0.0
sentencepiece>=0.1.99
deep-translator>=1.8.3
argostranslate>=1.8.0
